dpso_utils/windows_utils: Replace argvToCmdLine() with createCmdLine()
createCmdLine() is basically the same as argvToCmdLine(), but accepts the
program name as a separate argument. It turns out that CommandLineToArgv()
doesn't parse the program name (argv[0]) in the same way as the other
arguments: argv[0] can be double quoted, but other rules ("in quotes" mode,
backslash escapes, etc.) don't apply.