cmake/gen_manual: Use CMake's "copy" instead of "copy_if_different"
We don't need "copy_if_different" in that case.

If the source file is newer that the target, but its content it the same,
the source is not copied and the timestamp of the target doesn't change,
making the rule to be always out of date. This happens at least on Windows.

This behavior may be useful in some rare cases when it allows to avoid
expensive rebuilds, for example, if a bunch of targets depend on a file
that may be regenerated, but eventually end with the same content as the
existing one.

In our particular case that doesn't make sense.