dpso: Revise ProgressTracker usage
There's now a convention that a routine taking a progress tracker
always creates a local tracker under the hood. This way, the code at
upper levels don't have to make assumptions about the number of jobs;
start(), advanceJob(), and finish() are now always called from the
scope of the tracker.

The progress from dpsoOcrGetProgress() is now split between
prepareScreenshot() and OCR, while previously there were 3 jobs, as
prepareScreenshot() was flattened to resizing and unsharp masking.