ui/windows_rc: Add application manifest to enable UTF-8 code page
The related entry was removed from the Troubleshooting section of the
manual, since the program no longer loads data from the directory of
the executable. On Windows older than Windows 10 version 1903, the
problem may still arise when %LOCALAPPDATA% has characters outside the
current code page (e.g., in the user's name), but this is a very
unlikely case which is impossible to fix.