ui/ui_common: Revise directories API and add Windows support
Windows support is one of the reasons for revising. It turns out that
some Windows APIs (namely SHCreateItemFromParsingName()) don't accept
empty directory elements (e.g. two consecutive backslashes) that we
get when inserting uiDataDir. This can be fixed by using a dedicated
function for concatenating paths, but in general it's better to not
expose empty directory names and use a higher-level API instead.