dpso_utils/cfg: Remove special handling of double quotes
In version 1.0, any \ that doesn't form an escape sequence was kept as
is. In c9818614, we made escaping \ mandatory, which makes the special
handling of double quotes unnecessary: to preserve trailing spaces,
it's now possible to just put \ at the end of the line (or escape the
last space).

These changes are not backward-compatible, but the
action_copy_to_clipboard_text_separator option is currently the only
one that may have leading or trailing spaces or double quotes, so
let's hope we don't mess up things for most of the users.