ui/qt: Use icon theme only for app icon
The main app icon is the only icon we install in the "hicolor" theme.

This change is also necessary to make sure that the themed icon does
not appear in the tray, as it must match the "busy" and "error"
variants. Of course, we can provide a copy of "normal" icon especially
for the tray (e.g. "dpscreenocr-ready.png" or something similar), but
it probably doesn't make much practical sense since it's unlikely that
icon theme authors will draw all the custom variants of our icons.