CMake: Use *_STANDARD and instead of target_compile_features().
The reason is a bug in versions before 3.22: when the standard set
by target_compile_features() is less than or equal to the compiler
default, CMake ignores the disabled *_EXTENSIONS property and does
not add the necessary flag to disable language extensions (e.g. an
explicit -std=c++* instead of the implied -std=gnu++* in GCC).