Drop AppImage support in favor of a simple bundle
As of this writing, the main problem with AppImage is that the
universal runtime (the one with statically linked libfuse) is still
experimental and buggy, and the authors don't have enough time to fix
bugs and review pull requests. Without the universal runtime, we can
only use an old one that links against the system libfuse. Since
there are two major versions of libfuse, this boils down to either
requiring the user of legacy systems to install libfuse3, or
requiring users of modern systems to install libfuse2. Obviously,
this defeats the whole purpose of AppImage - to be user-friendly.

The second problem with AppImage is that you have to describe to the
user how to make it executable. A desktop environment could help here
(e.g. show a dialog to make an AppImage executable when the user
double-clicks on it), but none of DEs do this, and probably never will
since Snap and Flatpack seem to be more popular. In contrast, archives
don't have this usability problem: all DEs have nice GUIs for
unpacking archives, and the unpacked binaries will already have
executable bits.

The third is less a problem than a personal quibble. An AppImage is
basically a self-mounting Squashfs archive that is decompressed on the
fly while the program is in use. Even if the performance impact and
startup time isn't a big issue (i.e. you have an SSD, the program is
small, and Squashfs uses a fast compression algorithm like Zstd), I
still see absolutely no practical sense in doing these computations
after the program is already "installed" on the user's machine (BTW,
Snap and Flatpack do the same thing). I understand that this is a
consequence of the AppImage "no installation needed" philosophy, but
I personally think it would make more sense for Linux DEs to make a
proper installation framework like in macOS, which can already be
done trivially on top of AppImages by simply showing an installation
dialog when the user double-clicks an AppImage file.