Use fmt:: for string formatting
This is a part of a global goal to eliminate all locale-dependent
stuff. Using fmt gives more than 100 KB binary overhead, but this is
not a big issue since we will drop fmt in favor of <format> once we
have C++20 support.

There are still fprintf() calls to be removed where they are used for
file IO. We can't just blindly replace them with fmt::print() as it
throws std::system_error, so they should be rewritten to use exception
handling.