dpso_utils/os: Add locale-insensitive strerror() replacement
strerror() is one of the few cases where locale awareness is always a
problem. The error messages coming from our libraries are always in
English, and are not intended to be localized: they're supposed to be
sent be the user as part of a bug report. However, since the strerror
part is always gets translated, we - the developers - have to manually
translate it back to English to understand what's happened.