dpso_ext/history: Don't treat truncated data as error
This behavior is a disaster for the end user: if a history entry was
partially written (e.g., due to lack of disk space or power loss),
then it would be impossible to open the application the next time
without manually deleting the history file.

The new behavior is to load all correct entries and "restore" the
error state (we now also allow clearing the history while in this
state). It's not perfect, but still better than before. The next steps
would be:

* Instead of restoring the error state on opening, truncate the file
  to drop partially written entries and open as usual.

* Add a method to check if the history is in the error state, so that
  we can warn users before they try to add a new entry, rather than
  after the recently recognized text has already been lost because
  "Add to history" was the only active action.