Use custom formatting instead of fmt
We can't migrate to C++20 until 2032-04 when Ubuntu 22.04 reaches EOL
(<format> support was added in GCC 13, available in Ubuntu 24.04).
Until then, there is not much text formatting to justify the bloated
size of executables and tarballs. In fact, we have to stick with C++17
even longer, because our Linux bundle needs to support outdated
distributions (for now, the bundle is built with GCC 9, which is the
latest version available on Ubuntu 14.04 via an external PPA).

We already have the core format() implementation used in
dpsoStrNFormat(), and in most cases the migration boils down to
replacing "fmt::" with "str::". The new formatting routines have no
obvious disadvantages over <fmt> (at least in C++17 mode, where
compile-time checks of format strings are not available), except
probably the need for separate routines for justification and hex
conversion, but I personally find them much more readable than
complicated format specifications. In other words, there's a chance
that we will stick with this new custom formatting even once we have
<format> support.