ui/qt/main_window: Remove "Minimize on startup" checkbox
The ui_window_minimize_on_start setting was added in 0a8610e8 as a
workaround for those who want to add the program to autostart. It's
pretty useless on its own, without ui_window_minimize_to_tray, so it
doesn't make sense to expose it as a checkbox, especially now that we
have a proper "Run at system logon" setting.

We still keep this option in the CFG file for those who for some
reason want the program to be minimized to the tray on normal
invocation rather than on autostart.