dpso_utils/os_*: Use std::filesystem (Unix part)
Since we already use std::filesystem in ui_common/init_user_data.cpp,
it makes no sense to keep the old custom implementations of the OS
functions.

Currently, one test in test_os fails: GCC (but not Clang) says that
path::parent_path() of "//a" is "/" instead of "//". This shouldn't be
an issue in practice. We will see what to do with it once we implement
the Windows parts of this task.