dpso_utils/os_*: Use std::filesystem (Windows part)
There were two issues found in the GCC implementation of
std::filesystem on Windows:

* rename() reports errc::io_error on any error
  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=122726
* Path manipulation routines don't support UNC and device paths:
  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=99430

Fortunately, these bugs are not critical in our case.

getFileExt() was also rewritten using filesystem::path. The only
difference from the original implementation is that a trailing period
in a nonempty file name is now considered an extension. For example,
getFileExt("file.") now returns ".".